requirements.txt
Django==4.2.2
psycopg==3.1.9
django-mptt==0.14.0
django-ckeditor-5==0.2.8
django-jazzmin==2.6.0
django-analytical==3.1.0
django-taggit==4.0.0
Pillow==10.1.0
pytils~=0.4.1